requirements.txt
paho-mqtt==1.6.1
PyYAML==6.0
confp==0.4.1